US government data, 2003

Census data on the 50 states from 2003.

Format

A data frame with 50 observations on the following 24 variables.

State

the 50 states

Region

a factor with levels Midwest, Northeast, South, West

Pop

Population in 1000

Births

Number of births

Deaths

Number of deaths

Pop18

Percent of population 18 years of age or younger

Pop65

Percent of population 65 years of age or older

HSGrad

Percent of population 25 years of age or older with a high school degree

ColGrad

Percent of population 25 years of age or older with a college degree

TeacherPay

Average teachers salary in dollars

InfMortality

Infant mortality per 1000 live births

TeenBirths

Live births per 1000 15-19 year old females

ViolentCrime

Violent crime per 100000 population

PropertyCrime

Property crime per 100000 population

DeathPenalty

State has death penalty?

Executions

Number of executions 1977-2003

Poverty

Percent of populaton below the poverty level

Unemp

Percent unemployed (of population 16 years or older)

Uninsured

Percent uninsured (3 year aveage)

Income

Median household income in 1998 dollars

Earnings

Average hourly earnings of production workers in manufacturing

Heart

Deaths by heart disease per 100000 population

Vehicles

Deaths by motor vehicle accidents per 100000 population

Homeowners

Home ownership rate

Source

United States Census Bureau https://www.census.gov/